refactor: Replace `classes` modifier with helper
The modifier was overwriting the `classList` of an element, so no other classes could be provided by something like splattributes `...attributes`
