+++ packages/ember-core/package.json
@@ -228,9 +228,9 @@
       "./components/theme-switcher.js": "./dist/_app_/components/theme-switcher.js",
       "./components/toaster.js": "./dist/_app_/components/toaster.js",
       "./helpers/bind.js": "./dist/_app_/helpers/bind.js",
+      "./helpers/classes.js": "./dist/_app_/helpers/classes.js",
       "./helpers/index.js": "./dist/_app_/helpers/index.js",
       "./initializers/dayjs-plugins.js": "./dist/_app_/initializers/dayjs-plugins.js",
-      "./modifiers/classes.js": "./dist/_app_/modifiers/classes.js",
       "./modifiers/index.js": "./dist/_app_/modifiers/index.js",
       "./modifiers/on-click-outside.js": "./dist/_app_/modifiers/on-click-outside.js",
       "./modifiers/on-destroy.js": "./dist/_app_/modifiers/on-destroy.js",
