+++ packages/ember-core/src/helpers/index.ts
@@ -1,1 +1,2 @@
 export { bind, default as Bind } from './bind.ts';
+export { classes, default as Classes } from './classes.ts';
